pyproject.toml
[project]
name = "mcp-transcribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datetime>=5.5",
    "dotenv>=0.9.9",
    "mcp[cli]>=1.10.1",
    "notion-client>=2.4.0",
    "openai-whisper>=20250625",
    "requests>=2.32.4",
    "yt-dlp>=2025.6.30",
]
